+++ app/src/services/mailService.js
@@ -7,17 +7,10 @@ const CHANNEL = config.get('apiGateway.queueName');
 class MailService {
 
     constructor() {
-        logger.debug('Initializing queue with provider %s ', config.get('apiGateway.queueProvider'));
-        switch (config.get('apiGateway.queueProvider').toLowerCase()) {
-
-            case AsyncClient.REDIS:
-                this.asynClient = new AsyncClient(AsyncClient.REDIS, {
-                    url: config.get('apiGateway.queueUrl')
-                });
-                break;
-            default:
-
-        }
+        logger.debug('Initializing queue with provider Redis.');
+        this.asynClient = new AsyncClient(AsyncClient.REDIS, {
+            url: config.get('apiGateway.queueUrl')
+        });
         this.asynClient = this.asynClient.toChannel(CHANNEL);
     }